Corrections for compatibility with Docker
